moved all tests to the __tests__ directory
