+++ src/__tests__/components/FeedItem/FeedDescription.test.tsx
@@ -1,5 +1,5 @@
 import { shallow } from 'enzyme'
-import FeedDescription from '../FeedDescription'
+import FeedDescription from '../../../components/FeedItem/FeedDescription'
 import { AuthorMeta } from '../../../models/AuthorMeta'
 import { Hashtag } from '../../../models/Feed'
